suspense the table component
